Updated vehicle itinerary and vehicle attributes
